move db related files to under /yatai
